pyproject.toml
[tool.poetry]
name = "asanafs"
version = "0.1.0"
description = ""
authors = ["eli <eli@localhost>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fusepy = "^3.0.1"
asana = "^3.2.1"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
